WIP: refactoring out the abstract scheduler interface and add the DRF algorithm
