+++ src/ai/backend/manager/scheduler/__init__.py
@@ -1,0 +1,149 @@
+from __future__ import annotations
+
+from abc import ABCMeta, abstractmethod
+import logging
+from typing import (
+    Any, Optional,
+    Protocol,
+    Mapping,
+    Sequence,
+    Tuple,
+)
+import uuid
+
+from aiopg.sa.connection import SAConnection
+import attr
+
+from ai.backend.common.logging import BraceStyleAdapter
+from ai.backend.common.docker import (
+    ImageRef,
+)
+from ai.backend.common.types import (
+    AgentId, KernelId, AccessKey, SessionTypes,
+    ResourceSlot,
+)
+
+log = BraceStyleAdapter(logging.getLogger('ai.backend.manager.scheduler'))
+
+
+@attr.s(auto_attribs=True, slots=True)
+class AgentAllocationContext:
+    agent_id: AgentId
+    agent_addr: str
+    scaling_group: str
+
+
+@attr.s(auto_attribs=True, slots=True)
+class AgentContext:
+    agent_id: AgentId
+    agent_addr: str
+    scaling_group: str
+    available_slots: ResourceSlot
+    occupied_slots: ResourceSlot
+
+
+@attr.s(auto_attribs=True, slots=True)
+class ScheduleDecision:
+    agent_id: AgentId
+    kernel_id: KernelId
+
+
+@attr.s(auto_attribs=True, slots=True)
+class SchedulingContext:
+    '''
+    Context for each scheduling decision.
+    '''
+    registry: AgentRegistry
+    db_conn: SAConnection
+    known_slot_types: Mapping[str, str]
+
+
+@attr.s(auto_attribs=True, slots=True)
+class ExistingSession:
+    kernel_id: KernelId
+    access_key: AccessKey
+    sess_id: str
+    sess_type: SessionTypes
+    domain_name: str
+    group_id: uuid.UUID
+    scaling_group: str
+    image_ref: ImageRef
+    occupying_slots: ResourceSlot
+
+
+@attr.s(auto_attribs=True, slots=True)
+class PendingSession:
+    '''
+    Context for individual session-related information used during scheduling.
+    '''
+    kernel_id: KernelId
+    access_key: AccessKey
+    sess_id: str
+    sess_type: SessionTypes
+    domain_name: str
+    group_id: uuid.UUID
+    scaling_group: str
+    image_ref: ImageRef
+    resource_policy: dict
+    resource_opts: Mapping[str, Any]
+    requested_slots: ResourceSlot
+    target_sgroup_names: MutableSequence[str]
+    environ: Mapping[str, str]
+    mounts: Sequence[str]
+    mount_map: Mapping[str, str]
+    startup_command: Optional[str]
+    internal_data: Optional[Mapping[str, Any]]
+
+
+class PredicateCallback(Protocol):
+    async def __call__(self,
+                       sched_ctx: SchedulingContext,
+                       sess_ctx: SessionContext,
+                       db_conn: SAConnection = None) -> None:
+        ...
+
+
+@attr.s(auto_attribs=True, slots=True)
+class PredicateResult:
+    passed: bool
+    message: Optional[str] = None
+    success_cb: Optional[PredicateCallback] = None
+    failure_cb: Optional[PredicateCallback] = None
+
+
+class SchedulingPredicate(Protocol):
+    async def __call__(self,
+                       sched_ctx: SchedulingContext,
+                       sess_ctx: PendingSession) \
+                       -> PredicateResult:
+        ...
+
+
+class AbstractScheduler(metaclass=ABCMeta):
+
+    '''
+    Interface for scheduling algorithms where the
+    ``schedule()`` method is a pure function.
+    '''
+
+    config: Mapping[str, Any]
+
+    def __init__(self, config: Mapping[str, Any]) -> None:
+        self.config = config
+
+    @abstractmethod
+    def pick_session(
+            self,
+            total_capacity: ResourceSlot,
+            pending_sessions: Sequence[PendingSession],
+            existing_sessions: Sequence[ExistingSession],
+    ) -> Optional[KernelId]:
+        return None
+
+    @abstractmethod
+    def assign_agent(
+            self,
+            possible_agents: Sequence[AgentContext],
+            pending_session: PendingSession,
+    ) -> Optional[AgentId]:
+        return None
